# notes-backend/openapi.yaml
openapi: 3.0.3
info:
  title: 在线笔记平台 API
  version: 1.1.0
  description: |
    当前文档只描述已经实现并可调用的后端 HTTP API。
    - 统一响应包：`{ code, message, data, requestId, timestamp }`
    - 鉴权：`Authorization: Bearer <JWT>`
    - 未实现但有长期价值的接口记录在 `docs/api-contract-drift.md` 的 Planned APIs
servers:
  - url: http://localhost:3001
    description: 本地开发
...
paths:
...
  /api/ai/writer/stream:
    post:
      tags: [AI]
      summary: Stream AI writing text
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AiWriterInput'
      responses:
        '200':
          description: Plain text chunks streamed as server-sent text
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/ai/mindmap:
    post:
      tags: [AI]
      summary: Generate, expand, or optimize mind map data
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AiMindmapInput'
      responses:
        '201':
          $ref: '#/components/responses/AiMessagesEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/ai/mermaid:
    post:
      tags: [AI]
      summary: Generate Mermaid diagram code
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AiMermaidInput'
      responses:
        '201':
          $ref: '#/components/responses/AiMessagesEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/ai/knowledge-graph/proposal:
    post:
      tags: [AI]
      summary: Build a knowledge graph proposal for one knowledge base
      description: 手动触发单个知识库的图谱 proposal 构建；只返回建议，不写入核心图谱数据。
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AiKnowledgeGraphInput'
      responses:
        '201':
          $ref: '#/components/responses/AiKnowledgeGraphProposalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/ai/pet:
    post:
      tags: [AI]
      summary: Stream text-only AI pet chat
      description: The current provider route supports text chat only. Image chat returns 400.
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AiPetInput'
      responses:
        '200':
          description: Plain text chunks streamed as server-sent text
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/ai/summary:
    post:
      tags: [AI]
      summary: Generate an aggregate summary for selected notes
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/AiSummaryInput'
      responses:
        '201':
          $ref: '#/components/responses/AiSummaryEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/notes/{id}/versions:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Versions]
      summary: 版本列表
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    post:
      tags: [Versions]
      summary: 创建版本快照
      security:
        - BearerAuth: []
      responses:
        '201':
          $ref: '#/components/responses/Ok'
  /api/notes/{id}/versions/{versionNo}/restore:
    parameters:
      - $ref: '#/components/parameters/Id'
      - in: path
        name: versionNo
        required: true
        schema: { type: integer }
    post:
      tags: [Versions]
      summary: 恢复版本
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/tags:
    get:
      tags: [Tags]
      summary: 标签列表
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 标签列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tag'
    post:
      tags: [Tags]
      summary: 创建标签
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/TagInput'
      responses:
        '201':
          $ref: '#/components/responses/TagEnvelope'
  /api/tags/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Tags]
      summary: 更新标签
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/TagInput'
      responses:
        '200':
          $ref: '#/components/responses/TagEnvelope'
    delete:
      tags: [Tags]
      summary: 删除标签
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/tags/bulk:
    post:
      tags: [Tags]
      summary: 批量创建标签
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:
                  type: array
                  items: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/tags/merge:
    post:
      tags: [Tags]
      summary: 合并标签
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceIds, targetId]
              properties:
                sourceIds:
                  type: array
                  items: { type: string }
                targetId: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/tags/sync:
    post:
      tags: [Tags]
      summary: 同步标签计数
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/boards:
    post:
      tags: [Boards]
      summary: 创建画板
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/ResourceInput'
      responses:
        '201':
          $ref: '#/components/responses/ResourceEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/boards/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Boards]
      summary: 画板详情
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/ResourceEnvelope'
    put:
      tags: [Boards]
      summary: 更新画板
      security:
        - BearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/ResourceInput'
      responses:
        '200':
          $ref: '#/components/responses/ResourceEnvelope'
...
  /api/v1/semantic/search:
    get:
      tags: [Semantic]
      summary: 语义搜索
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: mode
          schema: { type: string, enum: [keyword, vector, hybrid], default: hybrid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CategoryId'
        - $ref: '#/components/parameters/TagIds'
      responses:
        '200':
          description: 语义搜索结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SemanticPage'
  /api/v1/semantic/topics:
    get:
      tags: [Semantic]
      summary: 主题列表
      security:
        - BearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/semantic/topics/convert:
    post:
      tags: [Semantic]
      summary: 主题转标签
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topicName, noteIds]
              properties:
                topicName: { type: string }
                noteIds:
                  type: array
                  items: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: notes_token
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
...
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100 }
...
    CategoryId:
      in: query
      name: categoryId
      schema: { type: string }
    TagIds:
      in: query
      name: tagIds
      schema:
        type: array
        items: { type: string }
      style: form
      explode: true
...
  requestBodies:
...
    TagInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Tag'
...
    ResourceInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceInput'
    AiWriterInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AiWriterInput'
    AiMindmapInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AiMindmapInput'
    AiMermaidInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AiMermaidInput'
    AiKnowledgeGraphInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AiKnowledgeGraphInput'
    AiPetInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AiPetInput'
    AiSummaryInput:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AiSummaryInput'
...
  responses:
    Ok:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
...
    TagEnvelope:
      description: 标签
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
...
    ResourceEnvelope:
      description: 资源
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Resource'
...
    AiMessagesEnvelope:
      description: AI legacy message wrapper
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AiMessagesOutput'
    AiSummaryEnvelope:
      description: AI summary result
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AiSummaryOutput'
    AiKnowledgeGraphProposalEnvelope:
      description: 知识图谱 proposal
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/KnowledgeGraphProposal'
...
    BadRequest:
      description: 请求错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
    NotFound:
      description: 不存在或无权访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiEnvelope'
  schemas:
    ApiEnvelope:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: OK }
        data: {}
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
...
    Tag:
      type: object
      properties:
        id: { type: string }
        _id: { type: string }
        name: { type: string }
        color: { type: string }
        noteCount: { type: integer }
...
    ResourceInput:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        noteId: { type: string }
        content: {}
    Resource:
      allOf:
        - $ref: '#/components/schemas/ResourceInput'
        - type: object
          properties:
            id: { type: string }
    SemanticPage:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        data:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              title: { type: string }
              preview: { type: string }
              score: { type: number }
              updatedAt: { type: string }
    AiWriterInput:
      type: object
      required: [context, type]
      properties:
        prompt: { type: string }
        context: { type: string }
        type: { type: string, enum: [continue, polish, summary] }
...
    AiMindmapInput:
      type: object
      required: [content]
      properties:
        content: {}
        scenario: { type: string, enum: [generate, expand, optimize], default: generate }
    AiMermaidInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        availableIcons:
          type: array
          items: { type: string }
    AiPetInput:
      type: object
      required: [message]
      properties:
        message: { type: string }
        conversationId: { type: string }
    AiSummaryInput:
      type: object
      required: [notes]
      properties:
        notes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AiSummaryNoteInput'
    AiSummaryNoteInput:
      type: object
      additionalProperties: true
      properties:
        title: { type: string }
        content: { type: string }
        updatedAt: { type: string, format: date-time }
    AiLegacyMessage:
      type: object
      required: [role, type, content]
      properties:
        role: { type: string, example: assistant }
        type: { type: string, example: answer }
        content: { type: string }
    AiMessagesOutput:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/AiLegacyMessage'
    AiSummaryOutput:
      type: object
      properties:
        summary: { type: string }
    AiKnowledgeGraphInput:
      type: object
      required: [knowledgeBaseId]
      properties:
        knowledgeBaseId: { type: string }
...
    KnowledgeGraphProposal:
      type: object
      properties:
        knowledgeBaseId: { type: string }
        generatedAt: { type: string, format: date-time }
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeGraphProposalNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeGraphProposalEdge'
        warnings:
          type: array
          items: { type: string }
    KnowledgeGraphProposalNode:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        type: { type: string, enum: [concept, entity, topic, claim] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        noteIds:
          type: array
          items: { type: string }
    KnowledgeGraphProposalEdge:
      type: object
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        relation: { type: string }
        weight: { type: number, minimum: 0, maximum: 1 }
        noteIds:
          type: array
          items: { type: string }
...