# notes-backend/src/modules/versions/versions.controller.ts
import { Controller, Get, Post, Param, Body, UseGuards, Request, Query } from '@nestjs/common'
import { AuthGuard } from '@nestjs/passport'
import { VersionsService } from './versions.service'

@UseGuards(AuthGuard('jwt'))
@Controller('notes/:id/versions')
export class VersionsController {
  constructor(private readonly service: VersionsService) { }

  @Get()
  async list(@Param('id') id: string, @Request() req) {
    return this.service.list(id, req.user.id)
  }

  @Post()
  async snapshot(@Param('id') id: string, @Body() body: any, @Request() req) {
    const rid = body?.requestId || (req.headers['x-request-id'] as string) || undefined
    return this.service.snapshot(id, req.user.id, body?.name, rid)
  }

  @Post(':versionNo/restore')
  async restore(@Param('id') id: string, @Param('versionNo') versionNo: string, @Body() body: any, @Request() req) {
    const rid = body?.requestId || (req.headers['x-request-id'] as string) || undefined
    return this.service.restore(id, Number(versionNo), req.user.id, rid)
  }
}