# notes-frontend/src/types/index.ts
// 用户相关类型
export interface User {
  id: string
  email: string
  displayName?: string
  avatarUrl?: string
  createdAt: string
  updatedAt: string
}

export interface LoginCredentials {
  email: string
  password: string
}

// 笔记相关类型
export interface NoteCategoryRef {
  id: string
  name: string
  color?: string
}

export interface NoteTagRef {
  id: string
  name: string
  color?: string
}

export interface Note {
  id: string
  title: string
  content: string
  summary?: string
  categoryId?: string
  category?: NoteCategoryRef | null
  tags: (string | NoteTagRef)[]
  createdAt: string
  updatedAt: string
  userId: string
  status?: 'published' | 'draft'
  visibility?: 'private' | 'public'
  acl?: Array<{ userId: string; role: 'owner' | 'editor' | 'viewer' }>
}

export interface CreateNoteDto {
  title: string
  content: string
  categoryId?: string
  tags: string[]
  status?: 'published' | 'draft'
  visibility?: 'private' | 'public'
}

export type UpdateNoteDto = Partial<CreateNoteDto>

export interface NoteFilterParams {
  keyword?: string
  categoryId?: string
  tagIds?: string[]
  tagsMode?: 'any' | 'all'
  startDate?: string
  endDate?: string
  status?: 'published' | 'draft'
  ids?: string[]
  searchMode?: 'text' | 'regex'
}

export interface KnowledgeBase {
  id: string
  name: string
  description: string
  createdAt: string
  updatedAt: string
}

export interface KnowledgeBaseNoteLink {
  id: string
  knowledgeBaseId: string
  noteId: string
  note: {
    id: string
    title: string
    summary?: string
    createdAt: string
    updatedAt: string
  }
  createdAt: string
}

export type KnowledgeGraphNodeType = 'concept' | 'entity' | 'topic' | 'claim'

export interface KnowledgeGraphNode {
  id: string
  label: string
  type: KnowledgeGraphNodeType
  confidence: number
  noteIds: string[]
}

export interface KnowledgeGraphEdge {
  id: string
  source: string
  target: string
  relation: string
  weight: number
  noteIds: string[]
}

export interface KnowledgeGraphProposal {
  knowledgeBaseId: string
  generatedAt: string
  nodes: KnowledgeGraphNode[]
  edges: KnowledgeGraphEdge[]
  warnings: string[]
}

export interface SavedFilter {
  id: string
  name: string
  criteria: NoteFilterParams
  userId: string
  createdAt: string
}

export interface CreateSavedFilterDto {
  name: string
  criteria: NoteFilterParams
}

// 分类相关类型
export interface Category {
  id: string
  name: string
  description?: string
  color?: string
  parentId?: string | null
  userId: string
  noteCount?: number
  createdAt: string
  updatedAt: string
}

// 标签相关类型
export interface Tag {
  id: string
  name: string
  color?: string
  noteCount?: number
  userId: string
  createdAt: string
}

export interface DashboardStats {
  notes: number
  categories: number
  tags: number
}

export interface RecentNoteSummary {
  id: string
  title: string
  preview: string
  updatedAt: string
  createdAt: string
  category?: NoteCategoryRef | null
  tags: NoteTagRef[]
}

export interface DashboardOverview {
  stats: DashboardStats
  recentNotes: RecentNoteSummary[]
  topCategories: Array<
    Pick<Category, 'id' | 'name' | 'color' | 'noteCount'>
  >
}